Seperate version command into its own file
