+++ peer/version/version.go
@@ -1,0 +1,48 @@
+/*
+Copyright IBM Corp. 2016 All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+		 http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package version
+
+import (
+	"fmt"
+
+	"github.com/hyperledger/fabric/core"
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+)
+
+// Cmd returns the Cobra Command for Version
+func Cmd() *cobra.Command {
+	return cobraCommand
+}
+
+var cobraCommand = &cobra.Command{
+	Use:   "version",
+	Short: "Print fabric peer version.",
+	Long:  `Print current version of fabric peer server.`,
+	PreRun: func(cmd *cobra.Command, args []string) {
+		core.LoggingInit("version")
+	},
+	Run: func(cmd *cobra.Command, args []string) {
+		Print()
+	},
+}
+
+// Print outputs the current executable version to stdout
+func Print() {
+	version := viper.GetString("peer.version")
+	fmt.Printf("Fabric peer server version %s\n", version)
+}
